+++ changelog.md
@@ -5,6 +5,7 @@
 - fixed: theme selector name/scrolling
 - fixed: dim on scroll
 - changed: disable autocomplete for search boxes
+- changed: filters extended
 
 ## Build 45